Can use Snarl without application password
[Application passwords](https://sites.google.com/site/snarlapp/developers/developer-guide#TOC-Application-Passwords) are used to prevent other applications from creating notifications by masquerading as an already registered application.

A Problem comes when using multiple instance of the same application.
In that case as we are generating a random password per instance, the second instance of an application cannot register itself while the first instance is running.
So client should be able to provide a password (always the same for a unique application) or no password at all.

See https://github.com/jcgay/maven-notifier/issues/13